Refactored to an INotificationHandler and skip inline tags in commands
